Add a build system to fix deploy cache invalidation issues and minify code.
