+++ .gitignore
@@ -1,1 +1,3 @@
-dist
+dist/
+public/
+node_modules/
