+++ src/404.html
@@ -7,10 +7,8 @@
 
     <title>Page not found · IsBGPSafeYet.com</title>
 
-    <link href="/favicon.ico?v=2" rel="icon" type="image/x-icon">
+    <link href="/resources/favicon.ico" rel="icon" type="image/x-icon">
     <link href="https://ui.components.workers.dev/?helpers=is-fullscreen,with-selection-styled,is-smooth-scrolling&components=Link,Button,Stack,Row,Markdown" rel="stylesheet">
-    <link rel="stylesheet" href="colors.css?v=21">
-    <link rel="stylesheet" href="index.css?v=21">
   </head>
 
   <body>
@@ -27,7 +25,5 @@
         </p>
       </div>
     </div>
-
-    <script src="https://unpkg.com/focus-visible-polyfill"></script>
   </body>
 </html>
