+++ src/css/index.css
@@ -1,3 +1,5 @@
+@import "colors.css";
+
 * {
   box-sizing: inherit;
 }
