+++ src/index.html
@@ -24,10 +24,8 @@
     <meta name="twitter:description" content="On the Internet, network devices exchange routes via a protocol called BGP (Border Gateway Protocol). Unfortunately, issues with BGP have led to malicious actors being able to hijack and misconfigure devices leading to security problems which have the potential to cause widespread problems. BGP security can be greatly improved by using technologies such as RPKI to sign Internet routes. This page attempts to track the progress of major Internet players (ISPs, transit operators, and content providers) in their progress to adopt RPKI and other technologies.">
     <meta name="twitter:url" content="https://isbgpsafeyet.com">
 
-    <link href="favicon.ico?v=2" rel="icon" type="image/x-icon">
+    <link href="/resources/favicon.ico" rel="icon" type="image/x-icon">
     <link href="https://ui.components.workers.dev/?helpers=with-selection-styled,is-smooth-scrolling&components=Link,Button,Stack,Row,Markdown" rel="stylesheet">
-    <link rel="stylesheet" href="colors.css?v=21">
-    <link rel="stylesheet" href="index.css?v=21">
   </head>
 
   <body>
@@ -356,10 +354,5 @@
 
       <div class="Spacer"></div>
     </div>
-
-    <script src="https://unpkg.com/focus-visible-polyfill"></script>
-
-    <script src="js/vendor/sortable.min.js"></script>
-    <script src="js/index.js?v=21"></script>
   </body>
 </html>
